# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Photo:
    properties:
      caption:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      userId:
        type: integer
    type: object
  entity.PhotoRequest:
    properties:
      caption:
        type: string
      photo_url:
        type: string
      title:
        type: string
    type: object
...
  entity.User:
    properties:
      age:
        type: integer
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  helper.BadRequest:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  helper.InternalServerError:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  helper.SuccessResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: febrianto.bekasi@gmail.com
    name: febrianto
  description: Final Project Digitalent x Hactive8
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mygram
  version: "1.0"
paths:
...
  /photo:
    get:
      consumes:
      - application/json
      description: Get all photos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Photo'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Get all photos
      tags:
      - photo
    post:
      consumes:
      - application/json
      description: Create new photo
      parameters:
      - description: create photo
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/entity.PhotoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Photo'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Create photo
      tags:
      - photo
  /photo/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the photo corresponding to the input Id
      parameters:
      - description: ID of the photo to be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Photo'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Delete photo identified by the given id
      tags:
      - photo
    get:
      consumes:
      - application/json
      description: Get details of photo corresponding is the input Id
      parameters:
      - description: ID of the photo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Photo'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Get photo details for a given id
      tags:
      - photo
    put:
      consumes:
      - application/json
      description: Update the photo corresponding to the input id
      parameters:
      - description: create photo
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/entity.PhotoRequest'
      - description: ID of the photo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Photo'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Update photo identified by the given id
      tags:
      - photo
...
schemes:
- http
securityDefinitions:
  JWT:
    description: 'How to input in swagger : ''Bearer <insert_your_token_here>'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"